+++ next.config.mjs
@@ -12,6 +12,9 @@ const nextConfig = {
         hostname: "via.placeholder.com"
       },
       {
+        hostname: "static.vecteezy.com"
+      },
+      {
         protocol: "https",
         hostname: "cdn-icons-png.flaticon.com",
         pathname: "/256/25/**"
