+++ src/app/globals.css
@@ -6,10 +6,10 @@
 @import "~slick-carousel/slick/slick-theme.css";
 
 body {
-  width: 100vw;
-  height: 100vh;
+  /* width: 100vw;
+  height: 100vh; */
   background-color: #54b2d3;
-  padding: 50px 130px 50px 50px;
+  /* padding: 50px 130px 50px 50px; */
 }
 
 button:hover {
