+++ src/components/NavigationBar.tsx
@@ -1,9 +1,8 @@
 import NavButton from "@/components/NavButton";
-import React from "react";
 
 const NavigationBar = () => {
   return (
-    <div className="flex flex-col gap-4 absolute right-[-85px] top-[140px] z-50">
+    <div className="flex flex-col gap-4 absolute right-[-100px] top-[150px] z-50">
       <NavButton>Home</NavButton>
       <NavButton>My Page</NavButton>
       <NavButton>Home</NavButton>
